updated SearchBar
