+++ Reci-Pie/ingredientstruct.swift
@@ -1,0 +1,14 @@
+//
+//  ingredientstruct.swift
+//  Reci-Pie
+//
+//  Created by Luis Morfin on 4/3/22.
+//
+
+import Foundation
+
+struct ingredientstruct:Decodable {
+    let title: String
+    let image: String
+    
+}
